pyproject.toml
[project]
name = "nanochat-desktop"
version = "0.1.0"
description = "Desktop AI chat application with NanoGPT"
requires-python = ">=3.11"
dependencies = [
    "gi>=3.0",
    "aiohttp>=3.9.0",
    "sqlalchemy>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["nanochat*"]

[project.scripts]
nanochat = "nanochat.main:main"
